Lab 4 7th Commit
Changes in full-volume-isotropic-phase.fs, which still needs to be improved.

+++ Part2/src/framework/light.h
@@ -6,7 +6,7 @@ class Light : public SceneNode {
 public:
 
 	float intensity;
-	float light_steps;
+	int light_steps;
 
 	Light(glm::vec3 position = glm::vec3(0.f), float intensity = 1.f, glm::vec4 color = glm::vec4(1.f));
